[PAY-2173] Make payments page work nicely in mobile web
