+++ packages/web/src/pages/pay-and-earn-page/PayAndEarnPage.module.css
@@ -5,6 +5,16 @@
   gap: var(--unit-8);
 }
 
+.mobilePageContainer {
+  user-select: none;
+  display: flex;
+  flex-direction: column;
+  gap: var(--unit-4);
+  padding: var(--unit-2);
+  margin-bottom: var(--unit-10);
+  padding-top: 64px;
+}
+
 .spinner {
   height: 40px;
   width: 40px;
