+++ packages/web/src/pages/pay-and-earn-page/components/WithdrawalsTab.tsx
@@ -20,17 +20,20 @@ import { useErrorPageOnFailedStatus } from 'hooks/useErrorPageOnFailedStatus'
 import { MainContentContext } from 'pages/MainContentContext'
 import { audiusBackendInstance } from 'services/audius-backend/audius-backend-instance'
 import { audiusSdk } from 'services/audius-sdk'
+import { isMobile } from 'utils/clientUtil'
 import { formatToday } from 'utils/dateUtils'
 import { useSelector } from 'utils/reducer'
 import { DASHBOARD_PAGE } from 'utils/route'
 
-import styles from './PayAndEarnPage.module.css'
-import { NoTransactionsContent } from './components/NoTransactionsContent'
+import styles from '../PayAndEarnPage.module.css'
+
+import { NoTransactionsContent } from './NoTransactionsContent'
 import {
   WithdrawalsTable,
+  WithdrawalsTableColumn,
   WithdrawalsTableSortDirection,
   WithdrawalsTableSortMethod
-} from './components/WithdrawalsTable'
+} from './WithdrawalsTable'
 
 const { getUserId } = accountSelectors
 
@@ -186,6 +189,10 @@ export const WithdrawalsTab = ({
 }: Omit<ReturnType<typeof useWithdrawals>, 'downloadCSV'>) => {
   const { mainContentRef } = useContext(MainContentContext)
 
+  const columns = isMobile()
+    ? (['date', 'amount'] as WithdrawalsTableColumn[])
+    : undefined
+
   return (
     <div className={styles.container}>
       {isEmpty ? (
@@ -193,6 +200,7 @@ export const WithdrawalsTab = ({
       ) : (
         <WithdrawalsTable
           key='withdrawals'
+          columns={columns}
           data={transactions}
           loading={isLoading}
           onSort={onSort}
