+++ packages/web/src/pages/remixes-page/components/mobile/RemixesPage.module.css
@@ -9,9 +9,6 @@
   margin-right: 4px;
 }
 
-.matrix {
-}
-
 .iconRemix path {
   fill: url(#linearGradient-1);
 }
